GCP/requirements.txt
google-api-python-client==2.56.0
google-auth==2.10.0
google-auth-httplib2==0.1.0
sendgrid==6.10.0